move marketplace to stocks controller, rename marketplace to markets
