+++ app/assets/stylesheets/application.sass.scss
@@ -4,7 +4,7 @@
 @import 'components/home';
 @import 'components/utility';
 @import 'components/devise';
-@import 'components/marketplace';
+@import 'components/markets';
 @import 'components/header';
 @import 'components/portfolio';
 @import 'components/transactions';
