--- app/helpers/marketplace_helper.rb
@@ -1,16 +1,0 @@
-module MarketplaceHelper
-
-  def format_date(date)
-    date.strftime("%m/%d/%Y")
-  end
-
-  def slice_text(text)
-    text.slice(0, 160)
-  end
-
-  def company_logo(symbol)
-    client = IEX::Api::Client.new
-    logo = client.logo(symbol)
-    logo.url
-  end
-end
