+++ app/views/layouts/application.html.erb
@@ -16,9 +16,9 @@
     <header class='app-header'>
       <%= image_tag("tradeX-logo", class: 'logo') %>
       <nav>
-        <%= link_to 'Marketplace', marketplace_index_path, class: 'link' %>
         <%= link_to 'My Portfolio', portfolio_path, class: 'link' %>
-        <a href="" class='link'>My Transactions</a>
+        <%= link_to 'My Transactions', transactions_path, class: 'link' %>
+        <%= link_to 'Markets', markets_path, class: 'link' %>
         <%= link_to 'Log Out', destroy_user_session_path, data: { turbo_method: :delete }, class:'link' %>
       </nav>
       <div class='profile'>
