+++ app/views/stocks/markets.html.erb
